Move selectors to platform
